+++ example/pprof/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"net/http/pprof"
+	"strings"
 
 	"github.com/teambition/gear"
 )
@@ -10,14 +11,12 @@ import (
 func main() {
 	app := gear.New()
 
-	router := gear.NewRouter()
 	// try: http://127.0.0.1:3000/debug/pprof
-	router.Get("/debug/pprof", gear.WrapHandlerFunc(pprof.Index))
-	router.Get("/debug/pprof/cmdline", gear.WrapHandlerFunc(pprof.Cmdline))
-	router.Get("/debug/pprof/profile", gear.WrapHandlerFunc(pprof.Profile))
-	router.Get("/debug/pprof/symbol", gear.WrapHandlerFunc(pprof.Symbol))
-	router.Get("/debug/pprof/trace", gear.WrapHandlerFunc(pprof.Trace))
-
-	app.UseHandler(router)
+	app.Use(func(ctx *gear.Context) error {
+		if strings.HasPrefix(ctx.Path, "/debug/pprof") {
+			pprof.Index(ctx.Res, ctx.Req)
+		}
+		return nil
+	})
 	app.Error(app.Listen(":3000"))
 }