+++ example/quic/app.go
@@ -45,7 +45,7 @@ func main() {}
 // 		return ctx.End(200, []byte(pushBody))
 // 	})
 // 	router.Get("/json", func(ctx *gear.Context) error {
-// 		return ctx.JSON(200, map[string]interface{}{"name": "quic"})
+// 		return ctx.JSON(200, map[string]any{"name": "quic"})
 // 	})
 // 	app.UseHandler(router)
 // 	app.Server.Addr = ":3000"