+++ util_test.go
@@ -26,22 +26,22 @@ import (
 
 // ----- Test Helpers -----
 
-func EqualPtr(t *testing.T, a, b interface{}) {
+func EqualPtr(t *testing.T, a, b any) {
 	assert.Equal(t, reflect.ValueOf(a).Pointer(), reflect.ValueOf(b).Pointer())
 }
 
-func NotEqualPtr(t *testing.T, a, b interface{}) {
+func NotEqualPtr(t *testing.T, a, b any) {
 	assert.NotEqual(t, reflect.ValueOf(a).Pointer(), reflect.ValueOf(b).Pointer())
 }
 
-func PickRes(res interface{}, err error) interface{} {
+func PickRes(res any, err error) any {
 	if err != nil {
 		panic(err)
 	}
 	return res
 }
 
-func PickError(res interface{}, err error) error {
+func PickError(res any, err error) error {
 	return err
 }
 
@@ -828,7 +828,7 @@ func TestGearValuesToStruct(t *testing.T) {
 		assert.NotNil(ValuesToStruct(data, v1, "form"))
 
 		v3 := struct {
-			String interface{} `form:"string"`
+			String any `form:"string"`
 		}{}
 		assert.NotNil(ValuesToStruct(data, &v3, "form"))